+++ CHANGELOG.md
@@ -38,7 +38,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ## [Unreleased]
 
 ### Improvements
-* (x/upgrade) [\#10532](https://github.com/cosmos/cosmos-sdk/pull/10532)  Update x/upgrade `keeper.DumpUpgradeInfoToDisk` to save `Plan.Info`.
+* (x/upgrade) [\#10532](https://github.com/cosmos/cosmos-sdk/pull/10532)  Add `keeper.DumpUpgradeInfoWithInfoToDisk` to include `Plan.Info` in the upgrade-info file.
 
 ### Bug Fixes